# JWT_spring_securtiy/src/main/java/ma/yasserprojects/jwt_spring_securtiy/auth/AuthenticationResponse.java
package ma.yasserprojects.jwt_spring_securtiy.auth;


import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;
}